requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pytest
azure-functions
azure-functions-durable
azure-storage-blob
pyodbc